Added silly message
When a user tries to calc a nutri value and get eg. 2, but the value is 0... we cannot get that value any greater than 0. No matter how big a number you multiply 0 by... it will still be 0. So we show a message using flasks flash function. "Silly! You can't get {desired_Value} from 0!"